move prediction api
