+++ inky-mbta-tracker/track_predictor/track_prediction_api.py
@@ -12,7 +12,8 @@ from mbta_client import determine_station_id
 from pydantic import BaseModel, ValidationError
 from shared_types.schema_versioner import schema_versioner
 from shared_types.shared_types import TrackAssignment, TrackPrediction
-from track_predictor import TrackPredictionStats, TrackPredictor
+
+from track_predictor.track_predictor import TrackPredictionStats, TrackPredictor
 
 # This is intended as a separate entrypoint to be run as a separate container
 
@@ -198,6 +199,4 @@ async def get_historical_assignments(
 @click.command()
 def run_main() -> None:
     port = int(os.environ.get("IMT_TRACK_API_PORT", "8080"))
-    with asyncio.Runner() as runner:
-        runner.run(schema_versioner())
-        uvicorn.run(app, host="0.0.0.0", port=port)
+    uvicorn.run(app, host="0.0.0.0", port=port)
